% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{close_browser}
\alias{close_browser}
\title{Close browser associated with RemoteDriver client}
\usage{
close_browser(driver)
}
\arguments{
\item{driver}{}
}
\value{

}
\description{
Close browser associated with RemoteDriver client
}
\examples{
driver <- open_browser("http://google.co.uk")
close_browser(driver)
}
