% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{Open_browser}
\alias{Open_browser}
\title{Open browser to URL}
\usage{
Open_browser(url = "")
}
\arguments{
\item{url}{URL to open}
}
\value{
RemoteDriver client
}
\description{
Open browser to URL
}
\examples{
driver <- open_browser("http://google.co.uk")
close_browser(driver)
}
